add lesson info api method and model
